requirements.txt
Django
django-filter
django-pytest
djangorestframework
mysqlclient
autopep8
django-cors-headers
pytz
